requirements.txt
streamlit>=1.32.0
folium>=0.16.0
streamlit-folium>=0.20.0
google-generativeai>=0.5.0
anthropic>=0.25.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
cohere>=5.0.0